Refactor patch_basic.py to enhance readability by replacing string concatenation with f-strings and improving error handling logic.

+++ modules_forge/patch_basic.py
@@ -25,7 +25,7 @@ def gradio_url_ok_fix(url: str) -> bool:
 
 
 def build_loaded(module, loader_name):
-    original_loader_name = loader_name + '_origin'
+    original_loader_name = f'{loader_name}_origin'
 
     if not hasattr(module, original_loader_name):
         setattr(module, original_loader_name, getattr(module, loader_name))
@@ -40,17 +40,16 @@ def build_loaded(module, loader_name):
             result = None
             exp = str(e) + '\n'
             for path in list(args) + list(kwargs.values()):
-                if isinstance(path, str):
+                if isinstance(path, str) and os.path.exists(path):
+                    exp += f'File corrupted: {path} \n'
+                    corrupted_backup_file = f'{path}.corrupted'
+                    if os.path.exists(corrupted_backup_file):
+                        os.remove(corrupted_backup_file)
+                    os.replace(path, corrupted_backup_file)
                     if os.path.exists(path):
-                        exp += f'File corrupted: {path} \n'
-                        corrupted_backup_file = path + '.corrupted'
-                        if os.path.exists(corrupted_backup_file):
-                            os.remove(corrupted_backup_file)
-                        os.replace(path, corrupted_backup_file)
-                        if os.path.exists(path):
-                            os.remove(path)
-                        exp += f'Forge has tried to move the corrupted file to {corrupted_backup_file} \n'
-                        exp += f'You may try again now and Forge will download models again. \n'
+                        os.remove(path)
+                    exp += f'Forge has tried to move the corrupted file to {corrupted_backup_file} \n'
+                    exp += f'You may try again now and Forge will download models again. \n'
             raise ValueError(exp)
         return result
